f: implement `KFoldValidation`
