+++ data/mapping/__init__.py
@@ -1,7 +1,10 @@
 from .noise import InstanceIndependantNoise  # noqa
 from .utils import (  # noqa
-    ClassBalanceTrainValSplit,
     PrecomputeDataset,
+)
+from .split import (  # noqa
+    ClassBalanceTrainValSplit,
+    KFoldValidation,
     RemapIndices,
     SubsetDataset,
     TrainValSplit,
